requirements.txt
pillow
pandas
sklearn
matplotlib
numpy
opencv-python
ipython 
jupyter
